Move self test component
